+++ src/systems/npc_system.cpp
@@ -1,10 +1,10 @@
 #include "npc_system.hpp"
 
 // Actualizar las IA
-void NPCSystem::update(EntityManager& em, float dt)
+void NPCSystem::update(EntityManager& em, double dt)
 {
-    em.forEach<SYSCMPs, SYSTAGs>([&,dt](Entity& e,PhysicsComponent& phy,RenderComponent& ren,NPCComponent &npc)
-    {   
+    em.forEach<SYSCMPs, SYSTAGs>([&, dt](Entity& e, PhysicsComponent& phy, RenderComponent& ren, NPCComponent& npc)
+    {
         NPCComponent* npcptr = &npc;
         if (npc.behaviourTree) {
             npc.behaviourTree->run({ em,e,nullptr,npcptr,phy,ren,nullptr,dt });
